Modify run.py to explore subdirectories
In order to group examples by topic it is needed to modify run.py in
order to find projects inside subfolders.

Currently, the qucsator test is not passing. HELP NEEDED!

Modify run.py so as to explore subdirectories

This commit is the first in a series in which the qucs examples will be
updated. As suggested at https://github.com/Qucs/qucs/issues/934 the
qucs examples set will be fetched from the qucs-test repo.

Since the qucs (test/example) projects will be grouped by topic, run.py
needs to be modified to as to find Qucs projects inside subdirectories