Modify testsuite/DC circuits/Operational amplifiers/General/DC_SP_opamp_gyrator_prj
Text overlapped with symbols and simulation boxes. It was added a text
block with the derivation of the input impedance of the gyrator.
Moreover, the input impedance is plotted in terms of resistance and
inductance. That proves that a 5H inductor was synthesized. Finally, the
feedback was changed to be negative feedback. The positive feedback is
highly likely to be unstable.