Modify testsuite/RFMW/Amplifiers/DC_AC_bbv_prj
The data display option was set not to open automatically. A graph was
added to show the gain. Some text overlapped with the symbols. It was
copied some theory from the product datasheet in the schematic since it
helps to understand the circuit.
The project name was renamed as Broadband_RFIC_NE5205_prj since it seems
to be more convenient.