Modify testsuite/RFMW/Filters/AC_SW_resonance_prj
The data display option was set not to open automatically. Some text
overlapped with the symbols. It were added two graphs showing the
voltage drop at the inductor and the capacitor as well as some math
formulae to explain how the voltage overshoot increases with Q in a
series resonator. The project was moved to
“testsuite/RFMW/Resonators/AC_SW_resonance_prj”.