Modify RFMW/General/AC_SW_swr_meter_prj
The data display option was set not to open automatically. Some text
overlapped with the symbols. Provided that both transformer and
resistors are broadband elements, the AC behavior is the same at 1GHz
and 2GHz. In favor of simplicity, the sweep was set to constant. A graph
showing the SWR (both measured and calculated) was added. The project
was moved to the ‘Couplers’ folder since it seems to be more convenient.