Modify testsuite/RFMW/Oscillators/TR_lc_osc_prj
The data display option was set not to open automatically. The project
was renamed as TR_Cross_Coupled_Oscillator_prj.  It was added an Octave
script to plot the spectrum. Using the series-to-parallel
transformation, the resistor was put in parallel with the L-C since it’s
more common to find that in the literature.