Modify testsuite/RFMW/Transmission Lines/SP_microstrip_prj
The data display option was set not to open automatically. Some text
overlapped with the symbols. The ‘Substrate’ property was hidden because
there’s only one substrate in the design. This may help to reduce the
clutter. It was added a graph to show the frequency response of the bias
tee. The project was renamed as SP_MicrostripBiasTee_prj since it seems
to be more appropiate.