Add three transistor current mirror project
This project was copied from
qucs.sourceforge.net/examples/Mirrors_prj.tar.gz. It comes with a
display sheet. For the sake of consistency with previous projects, the
graphs will be moved to the schematic sheet. The project is renamed as
DC_SW_ThreeTransistor_Current_Mirror_prj and moved to testsuite/DC
circuits/Current Mirrors