Add two stage amplifier project
This project was copied from
qucs.sourceforge.net/examples/2StageAmplifier_prj.tar.gz. The project
includes several schematics. In order to do the testing, the main
schematic (Amplifier.sch) was renamed as 2StageAmplifier and the project
whole project as SP_2StageAmplifier_prj. The data displays because they
are a lot. At the same time, this approach preserves the work of the
author.