Add UHF amplifier project. Author: Richard Crozier
This project was downloaded from
qucs.sourceforge.net/examples/UHF_Amplifier_prj.tar.gz. Some text
overlapped with the simulation box and the data display option was set
not to open automatically once the simulation ends. The project will be
skipped in the automatic test because the netlist generated during the
test includes the full path to the Qucs user folder in the S2P
component. This makes the test crash because the S2P is in the project
folder, not in the Qucs user folder (/home/username/.qucs).